bugfix-unprotected-routes

+++ autoscrape/routes.py
@@ -31,30 +31,31 @@ def login():
 				db.session.commit()
 
 		if current_user.is_authenticated:
-		    return redirect(url_for('dashboard'))
+				return redirect(url_for('dashboard'))
 		form = LoginForm()
 		print(form.username.data)
 		print(form.password.data)
 		# check if the form validated when it was submitted
 		if form.validate_on_submit():
-		    if (form.username.data == username) and (form.password.data == password): 
-		        login_user(user, remember=form.remember.data)
-		        flash(f"Logged in as {user.username}.","success")
-		        #use the "get()" dictionary method instead of [] so an error isnt thrown if the argument doesn't exist.
-		        next_page = request.args.get('next')
-		        return redirect(next_page) if next_page else redirect(url_for('dashboard'))
-		    else:
-		        flash('Login Unsuccessful. Please check username and password, and ensure AUTOSCRAPE_ADMIN_USERNAME and AUTOSCRAPE_ADMIN_PASSWORD environment variables have been set.', 'danger')
+				if (form.username.data == username) and (form.password.data == password): 
+						login_user(user, remember=form.remember.data)
+						flash(f"Logged in as {user.username}.","success")
+						#use the "get()" dictionary method instead of [] so an error isnt thrown if the argument doesn't exist.
+						next_page = request.args.get('next')
+						return redirect(next_page) if next_page else redirect(url_for('dashboard'))
+				else:
+						flash('Login Unsuccessful. Please check username and password, and ensure AUTOSCRAPE_ADMIN_USERNAME and AUTOSCRAPE_ADMIN_PASSWORD environment variables have been set.', 'danger')
 		return render_template("login.html", title="Login", form=form)
 
 @app.route("/logout")
 def logout():
-    logout_user()
-    return redirect(url_for('login'))
+		logout_user()
+		return redirect(url_for('login'))
 
 @app.route("/download/session_data/<int:session_id>")
 def download_session_data(session_id):
-	
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	data_entries = DataEntry.query.filter_by(session_id=session_id).order_by(DataEntry.timestamp.asc())
 	csv = db_query_output_to_csv.db_query_output_to_csv(
 		query_output=data_entries, 
@@ -67,6 +68,8 @@ def download_session_data(session_id):
 
 @app.route("/download/session_history")
 def download_session_history():
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	csv = db_query_output_to_csv(
 		query_output=Session.query.all(), 
 		columns_to_exclude=["_sa_instance_state", "id"])
@@ -79,6 +82,8 @@ def download_session_history():
 
 @app.route("/")
 def home():
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	return redirect(url_for("dashboard"))
 
 
@@ -106,7 +111,9 @@ def dashboard():
 
 
 @app.route("/session_data/<int:session_id>")
-def session_data(session_id):
+def session_data(session_id):	
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	session = Session.query.get_or_404(session_id)
 	data_entries = DataEntry.query.filter_by(session_id=session_id).order_by(DataEntry.timestamp.desc())
 	return render_template('session_data.html',
@@ -118,6 +125,8 @@ def session_data(session_id):
 
 @app.route("/log/<int:session_id>")
 def log(session_id):
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	session = Session.query.get_or_404(session_id)
 	log_entries = LogEntry.query.filter_by(session_id=session_id).order_by(LogEntry.date.desc())
 	data_entries = DataEntry.query.filter_by(session_id=session_id).order_by(DataEntry.timestamp.desc())
@@ -130,6 +139,8 @@ def log(session_id):
 
 @app.route("/create_session/<string:scraper_name>")
 def create_session(scraper_name):
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	Scraper = scrapers.get(scraper_name)
 	if Scraper:
 		if len(active_sessions) < max_active_sessions:
@@ -153,6 +164,8 @@ def create_session(scraper_name):
 
 @app.route("/abort_session/<string:session_id>")
 def abort_session(session_id):
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	try:
 		active_sessions[int(session_id)].destroy(completed=False)
 		flash(f"Scraper session {session_id} has been Aborted.","success")
@@ -163,6 +176,8 @@ def abort_session(session_id):
 
 @app.route("/delete_session_record/<string:session_id>")
 def delete_session_record(session_id):
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	try:
 		Session.query.filter_by(id=session_id).delete()
 		LogEntry.query.filter_by(session_id=session_id).delete()
@@ -176,6 +191,8 @@ def delete_session_record(session_id):
 
 @app.route("/abort_all_active_sessions")
 def abort_all_active_sessions():
+	if not current_user.is_authenticated:
+		return redirect(url_for('login'))
 	try:
 		#use shallow copy to break link back to active_sessions object
 		session_ids_to_destroy=[copy(session_id) for session_id in active_sessions.keys()]
